change to display-prettify; move display_messages to utils
