+++ parlai/scripts/interactive.py
@@ -24,7 +24,7 @@ def setup_args(parser=None):
     if parser is None:
         parser = ParlaiParser(True, True)
     parser.add_argument('-d', '--display-examples', type='bool', default=False)
-    parser.add_argument('--prettify', type='bool', default=False,
+    parser.add_argument('--display-prettify', type='bool', default=False,
                         help='Set to use a prettytable when displaying '
                              'examples with text candidates')
     return parser
