+++ projects/convai/convai_world.py
@@ -4,7 +4,8 @@
 # LICENSE file in the root directory of this source tree. An additional grant
 # of patent rights can be found in the PATENTS file in the same directory.
 
-from parlai.core.worlds import World, display_messages
+from parlai.core.worlds import World
+from parlai.core.utils import display_messages
 from parlai.core.agents import create_agent_from_shared
 
 import requests
